ai/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.0
SQLAlchemy==2.0.35
python-dotenv==1.0.1
alembic==1.13.2
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
Pillow==10.4.0
google-generativeai==0.8.3
python-multipart==0.0.9
ultralytics==8.0.196
opencv-python==4.8.1.78
easyocr==1.7.0
numpy==1.24.3
qdrant-client==1.9.2

